IS-1132 Reuse submit_retry
